fix: comprehensive bug sweep — 9 critical fixes across gameplay systems
- OverworldEnemy: startupClearCheckComplete no longer permanently skipped when tracker null
- QTESession: ShowQTE returns true on success; isComplete flag prevents coroutine race
- StatusEffect: Tick() preserves infinite (negative) duration effects
- QuestJournalUI: null guard on progression list; track DeactivateAfterDelay coroutine
- CombatSceneBootstrap: float cast in tier multiplier prevents integer division truncation
- AncientTextSceneBootstrap: null guard in GetViceSuffix for ResolveIslandId
- GameStateSerializer: CultureInfo.InvariantCulture for float JSON formatting
- TravelValidationService: filter unrestored islands from travel destinations
- SavePersistenceTest: null guards and IndexOf validation in fallback paths

+++ Assets/CombatSceneBootstrap.cs
@@ -877,7 +877,7 @@ public class CombatSceneBootstrap : MonoBehaviour
             if (string.Equals(progressionOrder[i], islandId, System.StringComparison.Ordinal))
             {
                 int tier = i + 1;
-                return Mathf.Lerp(1.0f, 1.65f, (tier - 1) / Mathf.Max(1, progressionOrder.Count - 1));
+                return Mathf.Lerp(1.0f, 1.65f, (float)(tier - 1) / Mathf.Max(1, progressionOrder.Count - 1));
             }
         }
